WIP: Moving vm-troubleshooter workflow to be under azure-diagnostics
