+++ plugin/skills/azure-compute/SKILL.md
@@ -1,6 +1,6 @@
 ---
 name: azure-compute
-description: "Azure VM/VMSS router. WHEN: create / provision / deploy / spin-up VM, recommend VM size, compare VM pricing, VMSS, scale set, autoscale, burstable, lightweight server, website, backend, GPU, machine learning, HPC simulation, dev/test, workload, family, load balancer, Flexible orchestration, Uniform orchestration, cost estimate, can't connect / RDP / SSH, refused, black screen, reset password, reach VM, port 3389, NSG, security, Linux, troubleshoot, troubleshooting, connectivity, capacity reservation (CRG), reserve, guarantee capacity, pre-provision, CRG association, CRG disassociation, machine enrollment (EMM), Essential Machine Management, monitor. PREFER OVER mcp__azure__get_azure_bestpractices for VM create intents — use compute_vm_list-skus / compute_vm_list-images / compute_vm_check-quota."
+description: "Azure VM/VMSS router. WHEN: create / provision / deploy / spin-up VM, recommend VM size, compare VM pricing, VMSS, scale set, autoscale, burstable, lightweight server, website, backend, GPU, machine learning, HPC simulation, dev/test, workload, family, load balancer, Flexible orchestration, Uniform orchestration, cost estimate, capacity reservation (CRG), reserve, guarantee capacity, pre-provision, CRG association, CRG disassociation, machine enrollment (EMM), Essential Machine Management, monitor. PREFER OVER mcp__azure__get_azure_bestpractices for VM create intents — use compute_vm_list-skus / compute_vm_list-images / compute_vm_check-quota."
 license: MIT
 metadata:
   author: Microsoft
@@ -15,7 +15,6 @@ Routes Azure VM and Virtual Machine Scale Set (VMSS) requests to the right workf
 
 - User wants to **recommend, compare, or price** a VM or VMSS
 - User wants to **create, provision, or deploy** a VM or VMSS
-- User **can't connect** to a VM (RDP / SSH / port refused / black screen / password reset)
 - User asks about **Capacity Reservation Groups** (CRG) — reserve, guarantee capacity, pre-provision
 - User asks about **Essential Machine Management** (EMM) — machine enrollment, monitor
 
@@ -27,7 +26,6 @@ Routes Azure VM and Virtual Machine Scale Set (VMSS) requests to the right workf
 Azure compute intent?
 ├── Recommend / compare / price a VM or VMSS         → VM Recommender
 ├── Create / provision / deploy a VM or VMSS         → VM Creator
-├── Can't connect / RDP / SSH / port refused         → VM Troubleshooter
 ├── Reserve / guarantee capacity (CRG)               → Capacity Reservation
 ├── Machine enrollment / Essential Machine Management → Essential Machine Management
 └── Unclear                                          → Ask which of the above
@@ -41,6 +39,5 @@ Azure compute intent?
 |---|---|---|
 | **VM Recommender** | [vm-recommender.md](workflows/vm-recommender/vm-recommender.md) | User asks which VM/VMSS to choose, wants pricing, or wants to compare options |
 | **VM Creator** | [vm-creator.md](workflows/vm-creator/vm-creator.md) | User wants to provision a bare VM or VMSS (not an app deployment) |
-| **VM Troubleshooter** | [vm-troubleshooter.md](workflows/vm-troubleshooter/vm-troubleshooter.md) | User can't connect, RDP/SSH refused, black screen, needs password reset |
 | **Capacity Reservation** | [capacity-reservation.md](workflows/capacity-reservation/capacity-reservation.md) | User needs to reserve / guarantee VM capacity (CRG create / associate / disassociate) |
 | **Essential Machine Management** | [essential-machine-management.md](workflows/essential-machine-management/essential-machine-management.md) | User asks about EMM / machine enrollment / monitor |
