+++ plugin/skills/azure-diagnostics/troubleshooting/compute/references/rdp-service-config.md
