+++ plugin/skills/azure-diagnostics/troubleshooting/compute/references/vm-agent-not-responding.md
